requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
openpyxl
pillow
